Refactored the SchemaGraph constructor into a more general one
