+++ graphql_compiler/tests/test_helpers.py
@@ -10,8 +10,10 @@ import six
 from .. import get_graphql_schema_from_orientdb_schema_data
 from ..debugging_utils import pretty_print_gremlin, pretty_print_match
 from ..schema_generation.schema_graph import SchemaGraph
-from ..schema_generation.utils import ORIENTDB_SCHEMA_RECORDS_QUERY
-
+from ..schema_generation.orientdb_schema_generation.utils import ORIENTDB_SCHEMA_RECORDS_QUERY
+from ..schema_generation.orientdb_schema_generation.orientdb_schema_graph import (
+    get_orientdb_schema_graph
+)
 
 # The strings which we will be comparing have newlines and spaces we'd like to get rid of,
 # so we can compare expected and produced emitted code irrespective of whitespace.
@@ -274,7 +276,7 @@ def generate_schema_graph(graph_client):
     """Generate SchemaGraph from a pyorient client"""
     schema_records = graph_client.command(ORIENTDB_SCHEMA_RECORDS_QUERY)
     schema_data = [x.oRecordData for x in schema_records]
-    return SchemaGraph(schema_data)
+    return get_orientdb_schema_graph(schema_data)
 
 
 def generate_schema(graph_client, class_to_field_type_overrides=None, hidden_classes=None):
