+++ graphql_compiler/tests/test_schema_generation.py
@@ -7,10 +7,12 @@ import six
 
 from .. import get_graphql_schema_from_orientdb_schema_data
 from ..schema_generation.graphql_schema import _get_union_type_name
-from ..schema_generation.schema_graph import SchemaGraph
-from ..schema_generation.schema_properties import (
+from ..schema_generation.orientdb_schema_generation.schema_properties import (
     ORIENTDB_BASE_EDGE_CLASS_NAME, ORIENTDB_BASE_VERTEX_CLASS_NAME, PROPERTY_TYPE_EMBEDDED_LIST_ID,
     PROPERTY_TYPE_EMBEDDED_SET_ID, PROPERTY_TYPE_LINK_ID, PROPERTY_TYPE_STRING_ID
+)
+from ..schema_generation.orientdb_schema_generation.orientdb_schema_graph import (
+    get_orientdb_schema_graph
 )
 
 
@@ -140,7 +142,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             BASE_VERTEX_SCHEMA_DATA,
             ENTITY_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertTrue(schema_graph.get_element_by_class_name('Entity').is_vertex)
 
     def test_parsed_edge(self):
@@ -152,17 +154,17 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             PERSON_LIVES_IN_EDGE_SCHEMA_DATA,
             PERSON_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertTrue(schema_graph.get_element_by_class_name('Person_LivesIn').is_edge)
 
     def test_parsed_non_graph_class(self):
         schema_data = [EXTERNAL_SOURCE_SCHEMA_DATA]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertTrue(schema_graph.get_element_by_class_name('ExternalSource').is_non_graph)
 
     def test_no_superclass(self):
         schema_data = [BASE_VERTEX_SCHEMA_DATA]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertEqual({ORIENTDB_BASE_VERTEX_CLASS_NAME},
                          schema_graph.get_inheritance_set(ORIENTDB_BASE_VERTEX_CLASS_NAME))
 
@@ -175,7 +177,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             PERSON_LIVES_IN_EDGE_SCHEMA_DATA,
             PERSON_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertEqual({'Person_LivesIn', ORIENTDB_BASE_EDGE_CLASS_NAME},
                          schema_graph.get_inheritance_set('Person_LivesIn'))
 
@@ -184,7 +186,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             BASE_VERTEX_SCHEMA_DATA,
             ENTITY_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         self.assertEqual({'Entity', ORIENTDB_BASE_VERTEX_CLASS_NAME},
                          schema_graph.get_inheritance_set('Entity'))
 
@@ -193,7 +195,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             BASE_VERTEX_SCHEMA_DATA,
             ENTITY_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         name_property = schema_graph.get_element_by_class_name('Entity').properties['name']
         self.assertTrue(name_property.type.is_same_type(GraphQLString))
 
@@ -203,7 +205,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             ENTITY_SCHEMA_DATA,
             PERSON_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         alias_property = schema_graph.get_element_by_class_name('Person').properties['alias']
         self.assertTrue(alias_property.type.is_same_type(GraphQLList(GraphQLString)))
         self.assertEqual(alias_property.default, set())
@@ -214,7 +216,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             DATA_POINT_SCHEMA_DATA,
             EXTERNAL_SOURCE_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         friends_property = schema_graph.get_element_by_class_name('DataPoint').properties[
             'data_source']
         self.assertTrue(friends_property.type.is_same_type(
@@ -230,7 +232,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             PERSON_LIVES_IN_EDGE_SCHEMA_DATA,
             PERSON_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         person_lives_in_edge = schema_graph.get_element_by_class_name('Person_LivesIn')
         self.assertEqual(person_lives_in_edge.base_in_connection, 'Person')
         self.assertEqual(person_lives_in_edge.base_out_connection, 'Location')
@@ -244,7 +246,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             PERSON_LIVES_IN_EDGE_SCHEMA_DATA,
             PERSON_SCHEMA_DATA,
         ]
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         person_lives_in_edge = schema_graph.get_element_by_class_name('Person_LivesIn')
         self.assertEqual(PERSON_LIVES_IN_EDGE_SCHEMA_DATA['customFields'],
                          person_lives_in_edge.class_fields)
@@ -262,7 +264,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
         schema, type_equivalence_dicts = get_graphql_schema_from_orientdb_schema_data(schema_data)
 
         # Sanity check
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         person_subclass_set = schema_graph.get_subclass_set('Person')
         self.assertIsNotNone(schema.get_type(_get_union_type_name(person_subclass_set)))
 
@@ -301,7 +303,7 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
         # from the type equivalence dict since it is not discoverable by the GraphQL Schema.
         self.assertEqual(0, len(type_equivalence_dicts))
         # Sanity check
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         person_subclass_set = schema_graph.get_subclass_set('Person')
         self.assertIsNone(schema.get_type(_get_union_type_name(person_subclass_set)))
 
@@ -317,6 +319,6 @@ class GraphqlSchemaGenerationTests(unittest.TestCase):
             PERSON_SCHEMA_DATA,
         ]
 
-        schema_graph = SchemaGraph(schema_data)
+        schema_graph = get_orientdb_schema_graph(schema_data)
         baby_lives_in_edge = schema_graph.get_element_by_class_name('Baby_LivesIn')
         self.assertEqual('Baby', baby_lives_in_edge.base_in_connection)
